+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1] - 2020-04-19
+
+### Changed
+
+- Update monkeymaster to 1.1.0 to cleanup storage.
+
 ## [1.2.0] - 2020-04-19
 
 ### Changed
@@ -36,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - command.games userscript and monkeymaster library
 - extends https://dominion.games/ to have a "Random!" button during card selection
 
-[unreleased]: https://github.com/davidtorosyan/command.games/compare/v1.2.0...HEAD
+[unreleased]: https://github.com/davidtorosyan/command.games/compare/v1.2.1...HEAD
+[1.2.1]: https://github.com/davidtorosyan/command.games/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/davidtorosyan/command.games/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/davidtorosyan/command.games/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/davidtorosyan/command.games/compare/v1.0.0...v1.1.0