+++ src/monkeymaster/CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2020-04-19
+
+### Changed
+
+- Modified how job data is stored and cleaned up. 
+  Now, by default, job data is deleted upon job completion.
+
+### Added
+
+- cleanupJobStorage API
+
 ## [1.0.0] - 2020-04-19
 
 ### Added
 
 - monkeymaster library
 
-[unreleased]: https://github.com/davidtorosyan/command.games/compare/monkeymaster-v1.0.0...HEAD
+[unreleased]: https://github.com/davidtorosyan/command.games/compare/monkeymaster-v1.1.0...HEAD
+[1.1.0]: https://github.com/davidtorosyan/command.games/releases/tag/monkeymaster-v1.1.0
 [1.0.0]: https://github.com/davidtorosyan/command.games/releases/tag/monkeymaster-v1.0.0